Refactor aggregator configuration (YML)
- custon NotNull validation (TODO: message interpolation)
- sane defaults, so configs can be smaller (less verbose)

+++ aggregator-core/src/main/java/eu/clarin/sru/fcs/aggregator/core/SearchGCParams.java
@@ -3,5 +3,5 @@ package eu.clarin.sru.fcs.aggregator.core;
 public interface SearchGCParams {
     int getSearchesSizeThreshold();
 
-    int getSearchesAgeThreshold();
+    long getSearchesAgeThreshold();
 }
